+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - `TeamCard.Image`
   - `TeamCard.PhoneIcon`
   - `TeamCard.Title`
-- Feature: Create `PhoneIcon` and `MailIcon` components.
+- Feature: Create `PhoneIcon` and `EmailIcon` components.
 - Fix: Fix `ExpandBoard.Button` arrow styles.
 - Fix: Add `box-sizing` CSS rule to `RewardCard` component.